+++ backend/Project/src/main/java/com/_thSem/Project/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package com._thSem.Project.repository;
 
-import com._thSem.Project.model.User;
+import com._thSem.Project.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
 }
